requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
Werkzeug==2.0.1
Flask-Session==0.4.0
bcrypt==3.2.0
SQLAlchemy==1.4.23
email-validator==1.1.3
flask-cors==5.0.1